bump version to 1.3 so that pypi will accept it.
basically this is necessary because i have made the mistake in the
past of using, for example, version numbers such as 1.02, which i
think is actually interpreted by pypi as 1.2.  laurie then "bumped"
this to 1.1.  but this is infact not bumping, it is actually putting
the version number backwards.  pypi then moans because we already have
a "1.01" (1.1) on pypi.

so basically, for example:

1.41 > 1.5   (in versioning number system)

if you make this mistake then it is a bit painful to have to fix.  so
do it right.  The former is minor revision fourty-one.  The latter is
minor revision five.  hence the former is actually more recent.